Script for checking assingments
This tool checks all the dirs (in a given dir) for .c files and compiles them.
After that, by a given set of scenarios, it checks the expected output of the program
and itself outputs the result of it.
The TODO's are so many, that I don't know where to begin.
Scenarios should be read from a file
The toplevel checking dir - also from a file (or ?)
Rethink the logic and see if a more fast (and little) version with more list comprehension is possible (really sleepy to think 'bout this right now)
Use Popen and communicate everywhere and remove os.system
...